# docs/phase0/openapi-lmn-authority-v1.yaml
openapi: 3.1.0
info:
  title: LMN Authority API
  version: 1.0.0
  summary: linuxmuster.net LINBO Host & Config Authority API
  description: |
    The LMN Authority API runs on the linuxmuster.net server (10.0.0.11) and
    serves as the **single source of truth** for LINBO host registrations,
    start.conf configurations, and DHCP reservations.

    **LINBO Docker consumes this API read-only.** Docker never writes host,
    config, or room data back to the LMN server. All CRUD for these entities
    happens on the LMN side (webui7 / linuxmuster-import-devices).

    The LINBO Runtime Docker stack polls this API via a cursor-based delta feed,
    fetches changed records in batches, and builds local snapshots that are
    served to PXE clients over TFTP / HTTP / rsync.

    ## Authentication
    Every request requires a **Bearer token** in the `Authorization` header.
    Additionally, the caller's IP address must be in the server's allowlist
    (typically `10.0.0.0/16`).

    ## Cursor Format
    The delta-feed cursor is an opaque string with the format
    `"<unix-timestamp>:<sequence>"`, e.g. `"1708943200:42"`.
    Cursors are monotonically increasing and safe to compare lexicographically.
    An empty or missing cursor triggers a full-snapshot response (all IDs).

    ## Rate Limits
    The API enforces a per-token rate limit of **60 requests / minute**.
    Batch endpoints count as a single request regardless of payload size.
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/linbo/startconfs:batch:
    post:
      operationId: batchGetStartConfs
      tags: [startconfs]
      summary: Batch fetch start.conf content by ID
      description: |
        Fetch the raw start.conf text for one or more config IDs. The content
        is the complete file as deployed to `/srv/linbo/start.conf.<id>`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    pattern: "^[a-zA-Z0-9_-]+$"
                  minItems: 1
                  maxItems: 100
            example:
              ids:
                - "win11_efi_sata"
                - "ubuntu_efi_nvme"
      responses:
        "200":
          description: start.conf records matching the requested IDs
          content:
            application/json:
              schema:
                type: object
                required: [startConfs]
                properties:
                  startConfs:
                    type: array
                    items:
                      $ref: "#/components/schemas/StartConfRecord"
              example:
                startConfs:
                  - id: "win11_efi_sata"
                    content: |
                      [LINBO]
                      Server = 10.0.0.1
                      Group = win11_efi_sata
                      Cache = /dev/sda4
                      SystemType = efi64
                      KernelOptions = quiet splash

                      [Partition]
                      Dev = /dev/sda1
                      Label = efi
                      Size = 200M
                      Id = ef
                      FSType = vfat
                      Bootable = yes

                      [Partition]
                      Dev = /dev/sda2
                      Label = msr
                      Size = 128M
                      Id = 0c01
                      FSType =
                      Bootable = no

                      [Partition]
                      Dev = /dev/sda3
                      Label = windows
                      Size = 80G
                      Id = 0700
                      FSType = ntfs
                      Bootable = no

                      [Partition]
                      Dev = /dev/sda4
                      Label = cache
                      Size =
                      Id = 8300
                      FSType = ext4
                      Bootable = no

                      [OS]
                      Name = Windows 11
                      Version = 24H2
                      IconName = win10.svg
                      BaseImage = win11.qcow2
                      Boot = /dev/sda3
                      Root = /dev/sda3
                      Kernel = auto
                      Initrd =
                      Append =
                      StartEnabled = yes
                      SyncEnabled = yes
                      NewEnabled = yes
                      Autostart = no
                      DefaultAction = sync
                      Hidden = no
                    hash: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
                    updatedAt: "2026-02-26T07:00:00Z"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"

  # =========================================================================
  # Batch Fetch — Configs (parsed)
  # =========================================================================
...
  /api/v1/linbo/startconf:
    get:
      operationId: getStartConf
      tags: [startconfs]
      summary: Look up a single start.conf by ID
      description: |
        Debug / admin endpoint. Not used in the boot-serving hot path.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]+$"
          example: "win11_efi_sata"
      responses:
        "200":
          description: start.conf found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartConfRecord"
              example:
                id: "win11_efi_sata"
                content: "[LINBO]\nServer = 10.0.0.1\nGroup = win11_efi_sata\n..."
                hash: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
                updatedAt: "2026-02-26T07:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: start.conf not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "NOT_FOUND"
                message: "No start.conf found with id 'nonexistent_config'"
        "429":
          $ref: "#/components/responses/RateLimited"

  # =========================================================================
  # DHCP Export — dnsmasq proxy
  # =========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: |
        Pre-shared API token configured on the linuxmuster.net server.
        Additionally, the caller's IP must be in the server's allowlist
        (default: `10.0.0.0/16`).

  # =========================================================================
  # Reusable Responses
  # =========================================================================
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "UNAUTHORIZED"
            message: "Missing or invalid Authorization header"
    Forbidden:
      description: Caller IP not in allowlist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "FORBIDDEN"
            message: "Source IP 192.168.1.50 is not in the allowlist"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "RATE_LIMITED"
            message: "Rate limit exceeded. Retry after 30 seconds."

  # =========================================================================
  # Schemas
  # =========================================================================
  schemas:
...
    StartConfRecord:
      type: object
      required: [id, content, hash, updatedAt]
      properties:
        id:
          type: string
          pattern: "^[a-zA-Z0-9_-]+$"
          description: Config name used as filename suffix (`start.conf.<id>`)
          examples: ["win11_efi_sata"]
        content:
          type: string
          description: Complete start.conf file content
        hash:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: SHA-256 hex digest of the content
          examples: ["a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"]
        updatedAt:
          type: string
          format: date-time
          examples: ["2026-02-26T07:00:00Z"]

    # -----------------------------------------------------------------------
    # Config (parsed)
    # -----------------------------------------------------------------------
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples: ["UNAUTHORIZED"]
        message:
          type: string
          description: Human-readable error description
          examples: ["Missing or invalid Authorization header"]
        details:
          type: object
          additionalProperties: true
          description: Optional structured error details